requirements.txt
# Core Framework
streamlit==1.32.0
streamlit-authenticator==0.3.2

# LLM & Orchestration
langchain==0.1.20
langchain-community==0.0.38
langchain-openai==0.1.6
langchain-core==0.1.52
openai==1.25.0

# Vector Database
chromadb==0.4.24

# Database
sqlalchemy>=2.0.36
alembic>=1.13.1

# Data Processing
pandas==2.2.1
numpy==1.26.4
python-dateutil==2.9.0

# Authentication & Security
bcrypt==4.1.2
python-jose==3.3.0
passlib==1.7.4

# Visualization
plotly==5.20.0
matplotlib==3.8.4

# Explainability
shap==0.45.0

# Utilities
python-dotenv==1.0.1
pydantic==2.7.0
pydantic-settings==2.2.1
requests==2.31.0
jinja2==3.1.3
fpdf2==2.7.9
openpyxl==3.1.2
pytz==2024.1
coloredlogs==15.0.1
tiktoken==0.6.0
sentence-transformers==2.7.0

# Alerting
smtplib2==0.2.1
